Revert web worker for coins data
