+++ src/redux/actions/coinsApiActions.js
@@ -2,7 +2,7 @@ import { CALL_API } from 'redux-api-middleware';
 import * as Actions from '../ActionNames';
 import config from 'config';
 
-export const fetchCoinsData = (locale, targetCurrency) => ({
+export const fetchCoinsData = (locale) => ({
   [CALL_API]: {
     endpoint: `${config.ORIGINS.COINS_IO}/front`,
     method: 'GET',
@@ -10,11 +10,7 @@ export const fetchCoinsData = (locale, targetCurrency) => ({
       Actions.FETCH_COINS_DATA,
       {
         type: Actions.FETCH_COINS_DATA_SUCCESS,
-        meta: {
-          locale,
-          targetCurrency,
-          WebWorker: true
-        }
+        meta: { locale }
       },
       Actions.FETCH_COINS_DATA_FAILURE
     ]
